+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* [App] Ability to toggle debug output
+  [#98](https://gitlab.data.bas.ac.uk/MAGIC/SIIS/-/issues/98)
+
 ## 1.0.0-rc.1 - 2021-07-15
 
 ### Changed [BREAKING!]